Revert "Replace Porpaginas with self-owned code"
This reverts commit 45fe74056f7a05424c279781625317e6f547e9a8.

This is required to keep compatibility with GraphQLite.
While this is technically a breaking change, the 6.0 version has low pulls and this interface is mostly internal.
The original commit was there to remove a deprecation warning. We still have no feedback from porpaginas project.
I believe that porpaginas will be updated once PHP effectively introduces the breaking change on the interface.